Change code to use a float instead of an int for the millicores_reserved value

+++ apps/main.go
@@ -136,11 +136,11 @@ func (a *Apps) GetUserID(username string) (string, error) {
 
 const setMillicoresStmt = `
 	UPDATE jobs
-	SET millicores_reserved = $2
+	SET millicores_reserved = $2::int
 	WHERE id = $1;
 `
 
-func (a *Apps) SetMillicoresReserved(analysisID string, millicores int) error {
+func (a *Apps) SetMillicoresReserved(analysisID string, millicores float64) error {
 	_, err := a.DB.Exec(setMillicoresStmt, analysisID, millicores)
 	return err
 }
